Fix paste issues (#3)
* Fix paste issues

This fixes a number of problems with paste/drop
- upload of image files contained in the text being pasted or dropped
- creates wiki links for links in the text being pasted or dropped
Note: This fix now bypasses the standard TinyMCE 'paste' plugin. This is because the 'paste' plugin was stripping out images from the text being pasted and dropped.  This is, hopefully a temporary measure and I will try using the 'pste' plugin again once upgraded to TinyMCE v 5

* Fix upload issues

When uploading a file there was a bug if mediawiki api responded with an error.  This change should fix that