Update 2020-06-11
Completed rewrite of conversion of html to wiki, including conversion of tables and lists. Rather than using regex on a serialised version of the html, the dome lelement s are processed before serialisation.
Also corrected problems with lists containing definition list elements.
Also now aligned with wikibase's code and updated to reflect patches posed in github for the existing version of the extension.