Update 2020-07-05
Changes to wikiparser to fix problem with loading event handlers when used on multiple fields withion PageForms.
Added functionality to $wgTinyMCESettings used in LocalSetting.php.  If a configuration variable is followed by a '-' and the correspnding value is '' then that variable will be removed from the default setting and the TinyMCE default value will be applied instead.  Documentation has been updated.