Update 20201109
This release includes changes to allow editing of 'plain text' blocks in the main editor eg <pre>, <nowiki> and pseudo pre blocks.

It allows editing of comments in the editor (by double clicking the comment placeholder).  There is also a menu option for inserting comments.

It also enable editing of references in the editor by double clickind the placeholder rather than haveing to toggle the vsibility mode and other refinements.

The button for editing raw wiki text has been changed to have a [[]] symbol rather than a 'bomb'.

I've incuded translation files for the extesnion that are available for the current version on mediawiki.