Release 1.03 2021-01-23
This release contains a number of fixes and new features as follows:
T6609 Conversion of dragged headers to wikitext headers on save
T6609 Removal of class=mw-headline if dragged header changed to diffrent element
T6596 processing of indentation inserted in TinyMCE editor
New features:
Wikilist plugin fixes a bug in the TinyMCE list plugin that meant style attributes werent saved when using tab and/or indent to add a new level to a list.

Functionality added to enable resizing buttons on toolbar. This uses the following config parameter in LocalSettings:
"toolbarResize" => "0.85"
Where 0.85 is the factor used to shrink the buttons.  A factor of 1 leaves the buttons at the normal size and greater than one makes them bigger.  Must be a real number greater than 0.

Added feature that the toolbar when editing inline is displayed abovbe the title on the viewed page rather than floating above the editable area.

Added a feature that shows a popup edit icon when the mouse is over an editable element on a viewed page.

Added a parser tag 'editinline' which can be used to bracket editable text on a viewed page.  When the page is parsed the content of the tags is placed in an editable div within a form element.  Currently, however, there is no means to save included in the extension!

Added config to MW_tinymce.js to display font and fontsize selector on the toolbar.

Included some additional langiage packs